test refinements
