+++ Tests/AudioKitTests/Node Tests/Generator Tests/AudioPlayer2Tests.swift
@@ -68,9 +68,12 @@ class AudioPlayer2Tests: XCTestCase {
         return url
     }
 
-    func createPlayer(duration: TimeInterval, buffered: Bool = false) -> AudioPlayer2? {
+    func createPlayer(duration: TimeInterval,
+                      frequencies: [AUValue]? = nil,
+                      buffered: Bool = false) -> AudioPlayer2? {
+        let frequencies = frequencies ?? chromaticScale
         guard let url = generateTestFile(ofDuration: duration,
-                                         frequencies: chromaticScale) else {
+                                         frequencies: frequencies) else {
             Log("Failed to open file")
             return nil
         }
@@ -146,181 +149,6 @@ extension AudioPlayer2Tests {
         player.play()
         XCTAssertTrue(player.isPlaying, "isPlaying should be true")
         player.stop()
-    }
-}
-
-// Real time functions, for local testing only
-extension AudioPlayer2Tests {
-    func testPause() {
-        realtimeTestPause()
-    }
-
-    func testScheduled() {
-        realtimeScheduleFile()
-    }
-
-    func testLooping() {
-        realtimeLoop(buffered: false)
-        realtimeLoop(buffered: true)
-    }
-
-    func testInterupts() {
-        realtimeInterrupts()
-    }
-
-    func realtimeTestPause() {
-        guard let player = createPlayer(duration: 6) else {
-            XCTFail("Failed to create AudioPlayer2")
-            return
-        }
-        let engine = AudioEngine()
-        engine.output = player
-        try? engine.start()
-
-        player.completionHandler = { Log("🏁 Completion Handler") }
-
-        var duration = player.duration
-
-        Log("▶️")
-        player.play()
-        wait(for: 2)
-        duration -= 2
-
-        Log("⏸")
-        player.pause()
-        wait(for: 1)
-        duration -= 1
-
-        Log("▶️")
-        player.play()
-        wait(for: duration)
-        Log("⏹")
-    }
-
-    func realtimeScheduleFile() {
-        guard let player = createPlayer(duration: 2) else {
-            XCTFail("Failed to create AudioPlayer2")
-            return
-        }
-        let engine = AudioEngine()
-        engine.output = player
-        try? engine.start()
-
-        var completionCounter = 0
-        player.completionHandler = {
-            completionCounter += 1
-            Log("🏁 Completion Handler", completionCounter)
-        }
-
-        // test schedule with play
-        player.play(at: AVAudioTime.now().offset(seconds: 3))
-
-        wait(for: player.duration + 3)
-
-        // test schedule separated from play
-        player.schedule(at: AVAudioTime.now().offset(seconds: 3))
-        player.play()
-
-        wait(for: player.duration + 3)
-
-        XCTAssertEqual(completionCounter, 2, "Completion handler wasn't called on both completions")
-    }
-
-    func realtimeLoop(buffered: Bool) {
-        guard let player = createPlayer(duration: 2, buffered: buffered) else {
-            XCTFail("Failed to create AudioPlayer2")
-            return
-        }
-        let engine = AudioEngine()
-        engine.output = player
-        try? engine.start()
-
-        var completionCounter = 0
-        player.completionHandler = {
-            if buffered {
-                XCTFail("For buffer looping the completion handler isn't called. The loop is indefinite")
-                return
-            }
-            completionCounter += 1
-            Log("🏁 Completion Handler", completionCounter)
-        }
-
-        player.isLooping = true
-        player.play()
-
-        wait(for: 5)
-        player.stop()
-    }
-
-    func realtimeInterrupts() {
-        guard let player = createPlayer(duration: 4, buffered: false) else {
-            XCTFail("Failed to create AudioPlayer2")
-            return
-        }
-        let engine = AudioEngine()
-        engine.output = player
-        try? engine.start()
-
-        player.play()
-        wait(for: 2)
-
-        guard let url2 = generateTestFile(ofDuration: 2,
-                                          frequencies: [220, 440]) else {
-            XCTFail("Failed to create file")
-            return
-        }
-
-        do {
-            let file = try AVAudioFile(forReading: url2)
-            try player.load(file: file)
-            XCTAssertNotNil(player.file, "File is nil")
-
-        } catch let error as NSError {
-            Log(error, type: .error)
-            XCTFail("Failed loading AVAudioFile")
-        }
-
-        wait(for: 1.5)
-
-        guard let url3 = generateTestFile(ofDuration: 3,
-                                          frequencies: [880, 220]) else {
-            XCTFail("Failed to create file")
-            return
-        }
-
-        do {
-            let file = try AVAudioFile(forReading: url3)
-            try player.load(file: file, buffered: true)
-            XCTAssertNotNil(player.buffer, "Buffer is nil")
-        } catch let error as NSError {
-            Log(error, type: .error)
-            XCTFail("Failed loading AVAudioFile")
-        }
-
-        wait(for: 2)
-
-        guard let url4 = generateTestFile(ofDuration: 3,
-                                          frequencies: chromaticScale),
-            let buffer = try? AVAudioPCMBuffer(url: url4) else {
-            XCTFail("Failed to create file or buffer")
-            return
-        }
-
-        player.buffer = buffer
-
-        wait(for: 1.5)
-
-        guard let url5 = generateTestFile(ofDuration: 1,
-                                          frequencies: chromaticScale.reversed()),
-            let file = try? AVAudioFile(forReading: url5) else {
-            XCTFail("Failed to create file or buffer")
-            return
-        }
-
-        player.file = file
-
-        wait(for: 2)
-        cleanup()
     }
 }
